move SplitPanel to file
